Only ingest messages from iframes we explicitly are watching
If more than one copy of iFrameResizer is loaded on the page, all
instances will receive the messages posted to the window. However they
should only process those messages if they came from an iframe that had
previously been passed through setupIFrame in this particular copy. So,
keep track of which iframes we are watching, and filter messages
accordingly.